Extract documentation related tasks from tasks.py
